refactor code follow n-tier
