+++ src/main/java/miu/waa/project1/controller/DiscussionController.java
@@ -3,8 +3,7 @@ package miu.waa.project1.controller;
 import lombok.RequiredArgsConstructor;
 import miu.waa.project1.dto.CreateOrUpdateDiscussionRequest;
 import miu.waa.project1.model.Discussion;
-import miu.waa.project1.model.DiscussionCategory;
-import miu.waa.project1.service.DiscussionService;
+import miu.waa.project1.service.impl.DiscussionServiceImpl;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,7 +13,7 @@ import java.util.List;
 @RequestMapping("/api/v1/discussion")
 @RequiredArgsConstructor
 public class DiscussionController {
-	private final DiscussionService discussionService;
+	private final DiscussionServiceImpl discussionService;
 
 	@GetMapping
 	public ResponseEntity<List<Discussion>> getAll() {
