+++ src/main/java/miu/waa/project1/controller/HealthController.java
@@ -1,20 +1,14 @@
 package miu.waa.project1.controller;
 
 import lombok.RequiredArgsConstructor;
-import miu.waa.project1.dto.CreateOrUpdateThreadRequest;
-import miu.waa.project1.model.Thread;
-import miu.waa.project1.service.ThreadService;
-import org.springframework.http.ResponseEntity;
+import miu.waa.project1.service.impl.ThreadServiceImpl;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("")
 @RequiredArgsConstructor
 public class HealthController {
-	private final ThreadService reviewService;
+	private final ThreadServiceImpl reviewService;
 
 	@GetMapping
 	public String check() {
