+++ src/main/java/miu/waa/project1/service/UserService.java
@@ -20,5 +20,5 @@ public interface UserService {
 
 	List<User> getAll();
 
-	boolean reportUser(Long id);
+	void reportUser(Long id);
 }
